Removed redundant casts
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk@2513 06c8f799-c779-4783-8af7-179afedb67c8

+++ classes/javax/microedition/lcdui/Graphics.java
@@ -242,7 +242,7 @@ public class Graphics
 
          // Calculate angle for partly circles / ellipses
          // NOTE: Below, (float) should not be needed. Not sure why Math.round() only accepts float.
-         int tp = (int) Math.round((float) Math.toDegrees(Math.atan2(yc, xc)));
+         int tp = (int) Math.round(Math.toDegrees(Math.atan2(yc, xc)));
          if (fill)
          {
             /* TODO: Optimize more by drawing horizontal lines */
